Restricting deletion of borrow record and allow deletion only if book's return status is 'true' and there is no lateReturnFee/ lateReturnFee is paid

+++ src/main/java/com/nischal/book_borrowing_app/service/BorrowService.java
@@ -1,5 +1,6 @@
 package com.nischal.book_borrowing_app.service;
 
+import com.nischal.book_borrowing_app.customError.CustomException;
 import com.nischal.book_borrowing_app.dto.BorrowResponseDTO;
 import com.nischal.book_borrowing_app.entity.Book;
 import com.nischal.book_borrowing_app.entity.Borrow;
@@ -35,7 +36,7 @@ public class BorrowService {
         Borrower borrower = borrowerRepository.findById(borrowerId).orElseThrow();
         // Check if the book can be borrowed
         if (book.getQuantity() <= 0 || !book.getAvailableStatus()) {
-            throw new RuntimeException("Book is not available for borrowing");
+            throw new CustomException("Conflict: Book is not available for borrowing");
         }
         if(borrower.getBooksBorrowed()==0)
         {
@@ -72,7 +73,13 @@ public class BorrowService {
     @Transactional
     public void deleteBorrow(Integer id) {
         Borrow borrow = borrowRepository.findById(id).orElseThrow();
-        borrowRepository.delete(borrow);
+        if (borrow.isReturnStatus() && borrow.getLateReturnFee()==0)
+        {
+            borrowRepository.delete(borrow);
+        }
+        else {
+            throw new CustomException("Conflict: Borrower has not returned the borrowed book or has pending late return fee");
+        }
     }
 
     public List<BorrowResponseDTO> getAllBorrows() {
